Chased around a special case that caused a small code reorg.
I had an image that when converted to G64 had an error 2 on a certain sector,
but when converting to D64 no such error occurred.  To troubleshoot, I
figured the problem must be in the track cycle detection somewhere cutting
off some bytes, but ultimately that wasn't it.  The sector header had some bad
GCR in it, but still passed the CBM XOR check.  The D64 conversion code does
not care and does not code error 6 (BAD GCR CODE) into the error block 
(I'm not even sure that's a valid error code in a D64, need to check).

When files are converted to G64, we always "fix" all bad GCR, which then
caused this block to fail the XOR and created an error 2 (header not found).

Weird situation.  This should be detected at disk read time, but I'd have
to change the code that converts to D64 sectors, as this same code is used
to check errors.  I think I should override on bad GCR, as this sector is
corrupt even if it passses the CBM XOR.

Anyway, the changes were to break out the track cycle alignment code out
of the read_nib function so it has to be called separately.  This way the D64
code doesn't need to operate on an already detected track cycle.  I may change
this back, but I like to break this out.