(diff from Nate Lawson) -----------------------
- Convert various parburst_read() and parburst_write() calls to the _n()
versions. This transfers a batch of reads or writes at once, which is
needed for good performance on USB (e.g., xum1541)

- Change the parallel port test routines to read data all in one batch
with cbm_parallel_burst_read_n().

- Change the send_mnib_cmd() API to take an optional array of byte args.
This is combined in a single call to cbm_parallel_burst_write_n().

- Add a 50 ms delay between running "UI" and then hard-reseting the bus.
Since there is no ack, it's possible to call them too quickly. This
delay may be a little long.

- Error out if unable to init the OpenCBM driver. This fixes a segfault
when an X-cable was not attached and the user called nibread/nibwrite.

- Use step_to_halftrack() instead of duplicating code.

- Resync WINDOWS/opencbm.h with the one from cvs