This is the first attempt  at getting rid of all the size_t errors that show up when building using the Windows 7 SDK.  A special thanks to Arnd Menge for pointing this out and including a Makefile and instructions for the SDK that exposed them all.  For whatever reason, the build process using the DDK does not show all the errors.
I did not attempt to remove "int" entirely from the program yet, but only the ones where it is compared and converted to other variables.  More work will be required to completely get rid of them, but I didn't want to completely break everything.

As it is, it exposed at least two bad (core dump) issues where we were working on data passed without checking if length>0, so this has been a very good exercise so far.  Thanks again Arnd.          

Of course thanks also to Nate Lawson who contributed the new USB support- I hope to convert soon and finally get rid of this parallel port, since I lose a few seconds on my Windows clock each time a disk is read or written. :)

Of course thanks to Spiro and Womo as always for continuing to develop OpenCBM!


Pete Rittwage