Fix reading of g64 from disk
speed entries can be pointer to speed blocks. to check for null, take
whole width of 4 bytes.

If track alignment is not the one given in the header and/or tracks are
not stored in sequential order (can happen if an emulator extends a given
image width further tracks), then we need to seek.
If we seek anyway, there is no need to read more then the actual track
size.