Fix GCR parsing.
On any g64 image created with kryoflux, end of sync is not aligned with
byte boundaries. So data has to be shifted before parsing.

The above fixes reading of most sectors.

One sector per track cannot be read after "shifting". It happens that
the end of track is between beginning of header (incl. sync) and end of
data.

So before parsing copy the length of the header block + data block into
a temporary buffer.